Calculate beta diversity dissimilarities with or without rarefaction

Calculate beta diversity dissimilarities with or without rarefaction

Usage

calcBeta(
  asvtable = NULL,
  metadata = NULL,
  numRare = 400,
  method = "bray",
  category = NULL,
  minReads = 1000,
  ...
)

Arguments

asvtable

Input ASV table. Can be filtered with filtering()

metadata

Metadata input object

numRare

Number of rarefactions to perform. Set to 0 if you do not want to do any rarefication.

method

Dissimilarity index to use. See ?vegan::vegdist for more information.

category

Name in metadata column that you'd like to use

minReads

Minimum number of reads to keep a sample

...

allow passed arguments to nested functions

Value

A dissimilarity matrix

Examples

asvtable <- unsampled_example()$asvtable
metadata <- unsampled_example()$metadata
calcBeta(asvtable = asvtable, metadata = metadata, category = "host_sex", method = "bray")
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#> Warning: function should be used for observed counts, but smallest count is 2
#>                                    SampleID         LabID   SampleType host_age
#> 1 5080-MS-1_328-GATCTACG-TCGACGAG_S328_L001 participant01 Homo sapiens       33
#> 2 5080-MS-1_339-ACTCACTG-GATCGTGT_S339_L001 participant02 Homo sapiens       25
#> 3 5348-MS-1_162-ACGTGCGC-GGATATCT_S162_L001 participant03 Homo sapiens       27
#> 4 5348-MS-1_297-GTCTGCTA-ACGTCTCG_S297_L001 participant04 Homo sapiens       26
#>   host_sex Host_disease   neg       PCoA1      PCoA2         PCoA3
#> 1   female      healthy FALSE  0.27230373  0.2565774 -0.3597532246
#> 2     male      healthy FALSE  0.06698387 -0.5154678 -0.0640682134
#> 3     male     COVID-19 FALSE -0.58088265  0.1190131  0.0001695812
#> 4   female     COVID-19 FALSE  0.24159504  0.1398773  0.4236518568
#> [1] TRUE
#> Error in rownames_to_column(x, var = id): Column name `SampleID` must not be duplicated.
#> Caused by error in `repaired_names()`:
#> ! Names must be unique.
#> ✖ These names are duplicated:
#>   * "SampleID" at locations 1 and 11.