Dada2 Wrapper

Dada2 Wrapper

Usage

dada2_wrapper(
  where = NULL,
  patternF = "_R1_001.fastq.gz",
  patternR = "_R2_001.fastq.gz",
  ...
)

Arguments

where

Path to your fastq files. Run dada2_wrapper(path = "example") if you want to run the example data.

patternF

Pattern of the forward/R1 reads

patternR

Pattern of the reverse/R2 reads

...

Can alter defaults of dada2 functions if needed

Value

dada2 ASV table and taxonomy table

Examples

dada2_wrapper(where = "example")
#> 
#> Please note that, because the main goal of dada2::filterAndTrim is to *write* filtered fastq files to the user's computer, and you chose to use the 'example' data, this function was NOT run. The filtered fastq reads already exist as example data for this package, so this package won't try to write the any data to your computer. With your own data, dada2::filterAndTrim will actually run, and make take some time, depending on the size of your input data!
#> 
#> 59520 total bases in 248 reads from 7 samples will be used for learning the error rates.
#> 49600 total bases in 248 reads from 7 samples will be used for learning the error rates.