Collection of helper/utility functions not really intended to be directly used by end user Path to user's fastq file folder

Collection of helper/utility functions not really intended to be directly used by end user Path to user's fastq file folder

Usage

whereFastqs(path = NULL, chatty = TRUE, return_tibble_or_path = "path")

Arguments

path

Path to folder containing fastq files

chatty

If set to FALSE, don't print as much to console.

return_tibble_or_path

Default is 'path', but instead can have it return a tibble of the files it found. should only ever be used for troubleshooting.

Value

file path to a folder

Examples

whereFastqs(".")
#> # A tibble: 35 × 1
#>    value              
#>    <chr>              
#>  1 assess_example.html
#>  2 assess_run-1.png   
#>  3 assess_run-2.png   
#>  4 assess_run-3.png   
#>  5 assess_run-4.png   
#>  6 assess_run-5.png   
#>  7 assess_run-6.png   
#>  8 assess_run-7.png   
#>  9 assess_run.html    
#> 10 calcBeta.html      
#> # ℹ 25 more rows
#> [1] "The total number of potential FASTQ files detected in the directory was 0, and the number of potential forward reads and reverse reads was 0. Please note that this is only performing simple pattern matching to look for standard Illumina-named files, and is only provided as a simple sanity check for you!"
#> [1] "."