Wrapper of entire dada2/decontam process

Wrapper of entire dada2/decontam process

Usage

dada2_decontam_wrapper(passedargs = FALSE, ...)

Arguments

passedargs

List passed arguments. Using only for troubleshooting during development; shouldn't have any use to end user

...

Arguments to pass to nested functions

Value

A list of the new ASV table, taxa table, and metadata object

Examples

dada2_decontam_wrapper(example = TRUE)
#> [1] "As no argument was provided for the 'path' of your fastq files, this wrapper will assume you want to work in your current directory, /home/runner/work/micro4R/micro4R/docs/reference"
#> [1] "Run with function dada2_asvtable started at 2025-10-03 16:48:09.356984"
#> # A tibble: 6 × 1
#>   value              
#>   <chr>              
#> 1 checkASV.html      
#> 2 checkMeta.html     
#> 3 contaminate.html   
#> 4 converter.html     
#> 5 dada2_asvtable.html
#> 6 index.html         
#> [1] "The total number of potential FASTQ files detected in the directory was 0, and the number of potential forward reads and reverse reads was 0. Please note that this is only performing simple pattern matching to look for standard Illumina-named files, and is only provided as a simple sanity check for you!"
#> Error in dada2_asvtable(example = FALSE, ...): No fastq files were found matching pattern '_R1_001.fastq.gz'! Check that 1) you're looking for the files the correct directory, and 2) you're looking for the right file name pattern.