Decontam

Decontam

Usage

decontam_wrapper(asvtable = NULL, taxa = NULL, metadata = NULL, ...)

Arguments

asvtable

asvtable

taxa

taxa

metadata

metadata

...

allow arguments to be passed to nested functions

Value

asvtable with potential contaminants removed

Examples

train <- "inst/extdata/db/EXAMPLE_silva_nr99_v138.2_toGenus_trainset.fa.gz"
species <- "inst/extdata/db/EXAMPLE_silva_v138.2_assignSpecies.fa.gz"

# the example ASV table is too small for decontam to work properly
# to get around this, we deliberately "contaminate" the ASV table with
# the contaminate() command. this requires us to use the matched metadata
# file created with contaminate()
asvtable <- converter(contaminate()$asvtable)
taxa <- dada2_taxa(asvtable = asvtable, train = train, species = species)
#> [1] "You're using the provided micro4R EXAMPLE reference databases. These are extremely tiny and unrealistic and meant only for testing and demonstrating purposes. DO NOT use them with your real data."
#> Finished processing reference fasta.
metadata <- contaminate()$metadata
decontam_wrapper(asvtable = asvtable, taxa = taxa, metadata = metadata, logfile = FALSE)
#> Warning: Removed 1 samples with zero total counts (or frequency).
#> Warning: Removed 1 samples with zero total counts (or frequency).
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 6 taxa and 10 samples ]
#> sample_data() Sample Data:       [ 10 samples by 6 sample variables ]
#> tax_table()   Taxonomy Table:    [ 6 taxa by 6 taxonomic ranks ]
#> $asvtable
#>                                                   TACGGAGGGTGCAAGCGTTAATCGGAATTACTGGGCGTAAAGCGCACGCAGGCGGTCTGTCAAGTCGGATGTGAAATCCCCGGGCTCAACCTGGGAACTGCATTCGAAACTGGCAGGCTAGAGTCTTGTAGAGGGGGGTAGAATTCCAGGTGTAGCGGTGAAATGCGTAGAGATCTGGAGGAATACCGGTGGCGAAGGCGGCCCCCTGGACAAAGACTGACGCTCAGGTGCGAAAGCGTGGGGAGCAAACAGG
#> SAMPLED_5080-MS-1_328-GATCTACG-TCGACGAG_S328_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5080-MS-1_339-ACTCACTG-GATCGTGT_S339_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5348-MS-1_162-ACGTGCGC-GGATATCT_S162_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5348-MS-1_297-GTCTGCTA-ACGTCTCG_S297_L001                                                                                                                                                                                                                                                            35
#> SAMPLED_5348-MS-1_381-TGCTCGTA-GTCAGATA_S381_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5080-MS-1_307-ATAGTACC-ACGTCTCG_S307_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5080-MS-1_313-GACATAGT-TCGACGAG_S313_L001                                                                                                                                                                                                                                                             0
#> FakeNeg1                                                                                                                                                                                                                                                                                                      0
#> FakeNeg2                                                                                                                                                                                                                                                                                                      0
#> FakeNeg3                                                                                                                                                                                                                                                                                                      0
#> FakeNeg4                                                                                                                                                                                                                                                                                                      0
#> FakeNeg5                                                                                                                                                                                                                                                                                                      0
#> FakeNeg6                                                                                                                                                                                                                                                                                                      0
#> FakeNeg7                                                                                                                                                                                                                                                                                                      0
#> FakeNeg8                                                                                                                                                                                                                                                                                                      0
#>                                                   TACGTAGGGTGCGAGCGTTGTCCGGAATTACTGGGCGTAAAGGGCTCGTAGGTGGTTTGTCGCGTCGTCTGTGAAATTCTGGGGCTTAACTCCGGGCGTGCAGGCGATACGGGCATAACTTGAGTGCTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCGAACAGG
#> SAMPLED_5080-MS-1_328-GATCTACG-TCGACGAG_S328_L001                                                                                                                                                                                                                                                              0
#> SAMPLED_5080-MS-1_339-ACTCACTG-GATCGTGT_S339_L001                                                                                                                                                                                                                                                              0
#> SAMPLED_5348-MS-1_162-ACGTGCGC-GGATATCT_S162_L001                                                                                                                                                                                                                                                             12
#> SAMPLED_5348-MS-1_297-GTCTGCTA-ACGTCTCG_S297_L001                                                                                                                                                                                                                                                              6
#> SAMPLED_5348-MS-1_381-TGCTCGTA-GTCAGATA_S381_L001                                                                                                                                                                                                                                                              0
#> SAMPLED_5080-MS-1_307-ATAGTACC-ACGTCTCG_S307_L001                                                                                                                                                                                                                                                              0
#> SAMPLED_5080-MS-1_313-GACATAGT-TCGACGAG_S313_L001                                                                                                                                                                                                                                                              0
#> FakeNeg1                                                                                                                                                                                                                                                                                                       0
#> FakeNeg2                                                                                                                                                                                                                                                                                                       0
#> FakeNeg3                                                                                                                                                                                                                                                                                                       0
#> FakeNeg4                                                                                                                                                                                                                                                                                                       0
#> FakeNeg5                                                                                                                                                                                                                                                                                                       0
#> FakeNeg6                                                                                                                                                                                                                                                                                                       0
#> FakeNeg7                                                                                                                                                                                                                                                                                                       0
#> FakeNeg8                                                                                                                                                                                                                                                                                                       0
#>                                                   TACGTAGGGTGCAAGCGTTGTCCGGAATTACTGGGCGTAAAGAGCTCGTAGGTGGTTTGTCACGTCGTCTGTGAAATTCCACAGCTTAACTGTGGGCGTGCAGGCGATACGGGCTGACTTGAGTACTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCAAACAGG
#> SAMPLED_5080-MS-1_328-GATCTACG-TCGACGAG_S328_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5080-MS-1_339-ACTCACTG-GATCGTGT_S339_L001                                                                                                                                                                                                                                                            13
#> SAMPLED_5348-MS-1_162-ACGTGCGC-GGATATCT_S162_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5348-MS-1_297-GTCTGCTA-ACGTCTCG_S297_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5348-MS-1_381-TGCTCGTA-GTCAGATA_S381_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5080-MS-1_307-ATAGTACC-ACGTCTCG_S307_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5080-MS-1_313-GACATAGT-TCGACGAG_S313_L001                                                                                                                                                                                                                                                             0
#> FakeNeg1                                                                                                                                                                                                                                                                                                      0
#> FakeNeg2                                                                                                                                                                                                                                                                                                      0
#> FakeNeg3                                                                                                                                                                                                                                                                                                      0
#> FakeNeg4                                                                                                                                                                                                                                                                                                      0
#> FakeNeg5                                                                                                                                                                                                                                                                                                      0
#> FakeNeg6                                                                                                                                                                                                                                                                                                      0
#> FakeNeg7                                                                                                                                                                                                                                                                                                      0
#> FakeNeg8                                                                                                                                                                                                                                                                                                      0
#>                                                   TACGTAGGTGACAAGCGTTGTCCGGATTTATTGGGCGTAAAGGGAGCGCAGGCGGTCTGTTTAGTCTAATGTGAAAGCCCACGGCTTAACCGTGGAACGGCATTGGAAACTGACAGACTTGAATGTAGAAGAGGAAAATGGAATTCCAAGTGTAGCGGTGGAATGCGTAGATATTTGGAGGAACACCAGTGGCGAAGGCGATTTTCTGGTCTAACATTGACGCTGAGGCTCGAAAGCGTGGGGAGCGAACAGG
#> SAMPLED_5080-MS-1_328-GATCTACG-TCGACGAG_S328_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5080-MS-1_339-ACTCACTG-GATCGTGT_S339_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5348-MS-1_162-ACGTGCGC-GGATATCT_S162_L001                                                                                                                                                                                                                                                            13
#> SAMPLED_5348-MS-1_297-GTCTGCTA-ACGTCTCG_S297_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5348-MS-1_381-TGCTCGTA-GTCAGATA_S381_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5080-MS-1_307-ATAGTACC-ACGTCTCG_S307_L001                                                                                                                                                                                                                                                             0
#> SAMPLED_5080-MS-1_313-GACATAGT-TCGACGAG_S313_L001                                                                                                                                                                                                                                                             0
#> FakeNeg1                                                                                                                                                                                                                                                                                                      0
#> FakeNeg2                                                                                                                                                                                                                                                                                                      0
#> FakeNeg3                                                                                                                                                                                                                                                                                                      0
#> FakeNeg4                                                                                                                                                                                                                                                                                                      0
#> FakeNeg5                                                                                                                                                                                                                                                                                                      0
#> FakeNeg6                                                                                                                                                                                                                                                                                                      0
#> FakeNeg7                                                                                                                                                                                                                                                                                                      0
#> FakeNeg8                                                                                                                                                                                                                                                                                                      0
#>                                                   TACGTAGGTCCCGAGCGTTGTCCGGATTTATTGGGCGTAAAGCGAGCGCAGGCGGTTAGATAAGTCTGAAGTTAAAGGCTGTGGCTTAACCATAGTACGCTTTGGAAACTGTTTAACTTGAGTGCAAGAGGGGAGAGTGGAATTCCATGTGTAGCGGTGAAATGCGTAGATATATGGAGGAACACCGGTGGCGAAAGCGGCTCTCTGGCTTGTAACTGACGCTGAGGCTCGAAAGCGTGGGGAGCAAACAGG
#> SAMPLED_5080-MS-1_328-GATCTACG-TCGACGAG_S328_L001                                                                                                                                                                                                                                                            0
#> SAMPLED_5080-MS-1_339-ACTCACTG-GATCGTGT_S339_L001                                                                                                                                                                                                                                                            0
#> SAMPLED_5348-MS-1_162-ACGTGCGC-GGATATCT_S162_L001                                                                                                                                                                                                                                                            3
#> SAMPLED_5348-MS-1_297-GTCTGCTA-ACGTCTCG_S297_L001                                                                                                                                                                                                                                                            0
#> SAMPLED_5348-MS-1_381-TGCTCGTA-GTCAGATA_S381_L001                                                                                                                                                                                                                                                            0
#> SAMPLED_5080-MS-1_307-ATAGTACC-ACGTCTCG_S307_L001                                                                                                                                                                                                                                                            0
#> SAMPLED_5080-MS-1_313-GACATAGT-TCGACGAG_S313_L001                                                                                                                                                                                                                                                            0
#> FakeNeg1                                                                                                                                                                                                                                                                                                     0
#> FakeNeg2                                                                                                                                                                                                                                                                                                     0
#> FakeNeg3                                                                                                                                                                                                                                                                                                     0
#> FakeNeg4                                                                                                                                                                                                                                                                                                     0
#> FakeNeg5                                                                                                                                                                                                                                                                                                     0
#> FakeNeg6                                                                                                                                                                                                                                                                                                     0
#> FakeNeg7                                                                                                                                                                                                                                                                                                     0
#> FakeNeg8                                                                                                                                                                                                                                                                                                     0
#> 
#> $taxa
#>                                                                                                                                                                                                                                                                 Kingdom
#> TACGGAGGGTGCAAGCGTTAATCGGAATTACTGGGCGTAAAGCGCACGCAGGCGGTCTGTCAAGTCGGATGTGAAATCCCCGGGCTCAACCTGGGAACTGCATTCGAAACTGGCAGGCTAGAGTCTTGTAGAGGGGGGTAGAATTCCAGGTGTAGCGGTGAAATGCGTAGAGATCTGGAGGAATACCGGTGGCGAAGGCGGCCCCCTGGACAAAGACTGACGCTCAGGTGCGAAAGCGTGGGGAGCAAACAGG  Bacteria
#> TACGTAGGGTGCGAGCGTTGTCCGGAATTACTGGGCGTAAAGGGCTCGTAGGTGGTTTGTCGCGTCGTCTGTGAAATTCTGGGGCTTAACTCCGGGCGTGCAGGCGATACGGGCATAACTTGAGTGCTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCGAACAGG Bacteria
#> TACGTAGGGTGCAAGCGTTGTCCGGAATTACTGGGCGTAAAGAGCTCGTAGGTGGTTTGTCACGTCGTCTGTGAAATTCCACAGCTTAACTGTGGGCGTGCAGGCGATACGGGCTGACTTGAGTACTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCAAACAGG  Bacteria
#> TACGTAGGTGACAAGCGTTGTCCGGATTTATTGGGCGTAAAGGGAGCGCAGGCGGTCTGTTTAGTCTAATGTGAAAGCCCACGGCTTAACCGTGGAACGGCATTGGAAACTGACAGACTTGAATGTAGAAGAGGAAAATGGAATTCCAAGTGTAGCGGTGGAATGCGTAGATATTTGGAGGAACACCAGTGGCGAAGGCGATTTTCTGGTCTAACATTGACGCTGAGGCTCGAAAGCGTGGGGAGCGAACAGG  Bacteria
#> TACGTAGGTCCCGAGCGTTGTCCGGATTTATTGGGCGTAAAGCGAGCGCAGGCGGTTAGATAAGTCTGAAGTTAAAGGCTGTGGCTTAACCATAGTACGCTTTGGAAACTGTTTAACTTGAGTGCAAGAGGGGAGAGTGGAATTCCATGTGTAGCGGTGAAATGCGTAGATATATGGAGGAACACCGGTGGCGAAAGCGGCTCTCTGGCTTGTAACTGACGCTGAGGCTCGAAAGCGTGGGGAGCAAACAGG   Bacteria
#>                                                                                                                                                                                                                                                                        Phylum
#> TACGGAGGGTGCAAGCGTTAATCGGAATTACTGGGCGTAAAGCGCACGCAGGCGGTCTGTCAAGTCGGATGTGAAATCCCCGGGCTCAACCTGGGAACTGCATTCGAAACTGGCAGGCTAGAGTCTTGTAGAGGGGGGTAGAATTCCAGGTGTAGCGGTGAAATGCGTAGAGATCTGGAGGAATACCGGTGGCGAAGGCGGCCCCCTGGACAAAGACTGACGCTCAGGTGCGAAAGCGTGGGGAGCAAACAGG  Pseudomonadota
#> TACGTAGGGTGCGAGCGTTGTCCGGAATTACTGGGCGTAAAGGGCTCGTAGGTGGTTTGTCGCGTCGTCTGTGAAATTCTGGGGCTTAACTCCGGGCGTGCAGGCGATACGGGCATAACTTGAGTGCTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCGAACAGG Actinomycetota
#> TACGTAGGGTGCAAGCGTTGTCCGGAATTACTGGGCGTAAAGAGCTCGTAGGTGGTTTGTCACGTCGTCTGTGAAATTCCACAGCTTAACTGTGGGCGTGCAGGCGATACGGGCTGACTTGAGTACTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCAAACAGG  Actinomycetota
#> TACGTAGGTGACAAGCGTTGTCCGGATTTATTGGGCGTAAAGGGAGCGCAGGCGGTCTGTTTAGTCTAATGTGAAAGCCCACGGCTTAACCGTGGAACGGCATTGGAAACTGACAGACTTGAATGTAGAAGAGGAAAATGGAATTCCAAGTGTAGCGGTGGAATGCGTAGATATTTGGAGGAACACCAGTGGCGAAGGCGATTTTCTGGTCTAACATTGACGCTGAGGCTCGAAAGCGTGGGGAGCGAACAGG       Bacillota
#> TACGTAGGTCCCGAGCGTTGTCCGGATTTATTGGGCGTAAAGCGAGCGCAGGCGGTTAGATAAGTCTGAAGTTAAAGGCTGTGGCTTAACCATAGTACGCTTTGGAAACTGTTTAACTTGAGTGCAAGAGGGGAGAGTGGAATTCCATGTGTAGCGGTGAAATGCGTAGATATATGGAGGAACACCGGTGGCGAAAGCGGCTCTCTGGCTTGTAACTGACGCTGAGGCTCGAAAGCGTGGGGAGCAAACAGG        Bacillota
#>                                                                                                                                                                                                                                                                              Class
#> TACGGAGGGTGCAAGCGTTAATCGGAATTACTGGGCGTAAAGCGCACGCAGGCGGTCTGTCAAGTCGGATGTGAAATCCCCGGGCTCAACCTGGGAACTGCATTCGAAACTGGCAGGCTAGAGTCTTGTAGAGGGGGGTAGAATTCCAGGTGTAGCGGTGAAATGCGTAGAGATCTGGAGGAATACCGGTGGCGAAGGCGGCCCCCTGGACAAAGACTGACGCTCAGGTGCGAAAGCGTGGGGAGCAAACAGG  Gammaproteobacteria
#> TACGTAGGGTGCGAGCGTTGTCCGGAATTACTGGGCGTAAAGGGCTCGTAGGTGGTTTGTCGCGTCGTCTGTGAAATTCTGGGGCTTAACTCCGGGCGTGCAGGCGATACGGGCATAACTTGAGTGCTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCGAACAGG      Actinobacteria
#> TACGTAGGGTGCAAGCGTTGTCCGGAATTACTGGGCGTAAAGAGCTCGTAGGTGGTTTGTCACGTCGTCTGTGAAATTCCACAGCTTAACTGTGGGCGTGCAGGCGATACGGGCTGACTTGAGTACTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCAAACAGG       Actinobacteria
#> TACGTAGGTGACAAGCGTTGTCCGGATTTATTGGGCGTAAAGGGAGCGCAGGCGGTCTGTTTAGTCTAATGTGAAAGCCCACGGCTTAACCGTGGAACGGCATTGGAAACTGACAGACTTGAATGTAGAAGAGGAAAATGGAATTCCAAGTGTAGCGGTGGAATGCGTAGATATTTGGAGGAACACCAGTGGCGAAGGCGATTTTCTGGTCTAACATTGACGCTGAGGCTCGAAAGCGTGGGGAGCGAACAGG              Bacilli
#> TACGTAGGTCCCGAGCGTTGTCCGGATTTATTGGGCGTAAAGCGAGCGCAGGCGGTTAGATAAGTCTGAAGTTAAAGGCTGTGGCTTAACCATAGTACGCTTTGGAAACTGTTTAACTTGAGTGCAAGAGGGGAGAGTGGAATTCCATGTGTAGCGGTGAAATGCGTAGATATATGGAGGAACACCGGTGGCGAAAGCGGCTCTCTGGCTTGTAACTGACGCTGAGGCTCGAAAGCGTGGGGAGCAAACAGG               Bacilli
#>                                                                                                                                                                                                                                                                           Order
#> TACGGAGGGTGCAAGCGTTAATCGGAATTACTGGGCGTAAAGCGCACGCAGGCGGTCTGTCAAGTCGGATGTGAAATCCCCGGGCTCAACCTGGGAACTGCATTCGAAACTGGCAGGCTAGAGTCTTGTAGAGGGGGGTAGAATTCCAGGTGTAGCGGTGAAATGCGTAGAGATCTGGAGGAATACCGGTGGCGAAGGCGGCCCCCTGGACAAAGACTGACGCTCAGGTGCGAAAGCGTGGGGAGCAAACAGG  Enterobacterales
#> TACGTAGGGTGCGAGCGTTGTCCGGAATTACTGGGCGTAAAGGGCTCGTAGGTGGTTTGTCGCGTCGTCTGTGAAATTCTGGGGCTTAACTCCGGGCGTGCAGGCGATACGGGCATAACTTGAGTGCTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCGAACAGG  Mycobacteriales
#> TACGTAGGGTGCAAGCGTTGTCCGGAATTACTGGGCGTAAAGAGCTCGTAGGTGGTTTGTCACGTCGTCTGTGAAATTCCACAGCTTAACTGTGGGCGTGCAGGCGATACGGGCTGACTTGAGTACTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCAAACAGG   Mycobacteriales
#> TACGTAGGTGACAAGCGTTGTCCGGATTTATTGGGCGTAAAGGGAGCGCAGGCGGTCTGTTTAGTCTAATGTGAAAGCCCACGGCTTAACCGTGGAACGGCATTGGAAACTGACAGACTTGAATGTAGAAGAGGAAAATGGAATTCCAAGTGTAGCGGTGGAATGCGTAGATATTTGGAGGAACACCAGTGGCGAAGGCGATTTTCTGGTCTAACATTGACGCTGAGGCTCGAAAGCGTGGGGAGCGAACAGG   Lactobacillales
#> TACGTAGGTCCCGAGCGTTGTCCGGATTTATTGGGCGTAAAGCGAGCGCAGGCGGTTAGATAAGTCTGAAGTTAAAGGCTGTGGCTTAACCATAGTACGCTTTGGAAACTGTTTAACTTGAGTGCAAGAGGGGAGAGTGGAATTCCATGTGTAGCGGTGAAATGCGTAGATATATGGAGGAACACCGGTGGCGAAAGCGGCTCTCTGGCTTGTAACTGACGCTGAGGCTCGAAAGCGTGGGGAGCAAACAGG    Lactobacillales
#>                                                                                                                                                                                                                                                                            Family
#> TACGGAGGGTGCAAGCGTTAATCGGAATTACTGGGCGTAAAGCGCACGCAGGCGGTCTGTCAAGTCGGATGTGAAATCCCCGGGCTCAACCTGGGAACTGCATTCGAAACTGGCAGGCTAGAGTCTTGTAGAGGGGGGTAGAATTCCAGGTGTAGCGGTGAAATGCGTAGAGATCTGGAGGAATACCGGTGGCGAAGGCGGCCCCCTGGACAAAGACTGACGCTCAGGTGCGAAAGCGTGGGGAGCAAACAGG  Enterobacteriaceae
#> TACGTAGGGTGCGAGCGTTGTCCGGAATTACTGGGCGTAAAGGGCTCGTAGGTGGTTTGTCGCGTCGTCTGTGAAATTCTGGGGCTTAACTCCGGGCGTGCAGGCGATACGGGCATAACTTGAGTGCTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCGAACAGG Corynebacteriaceae
#> TACGTAGGGTGCAAGCGTTGTCCGGAATTACTGGGCGTAAAGAGCTCGTAGGTGGTTTGTCACGTCGTCTGTGAAATTCCACAGCTTAACTGTGGGCGTGCAGGCGATACGGGCTGACTTGAGTACTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCAAACAGG  Corynebacteriaceae
#> TACGTAGGTGACAAGCGTTGTCCGGATTTATTGGGCGTAAAGGGAGCGCAGGCGGTCTGTTTAGTCTAATGTGAAAGCCCACGGCTTAACCGTGGAACGGCATTGGAAACTGACAGACTTGAATGTAGAAGAGGAAAATGGAATTCCAAGTGTAGCGGTGGAATGCGTAGATATTTGGAGGAACACCAGTGGCGAAGGCGATTTTCTGGTCTAACATTGACGCTGAGGCTCGAAAGCGTGGGGAGCGAACAGG   Carnobacteriaceae
#> TACGTAGGTCCCGAGCGTTGTCCGGATTTATTGGGCGTAAAGCGAGCGCAGGCGGTTAGATAAGTCTGAAGTTAAAGGCTGTGGCTTAACCATAGTACGCTTTGGAAACTGTTTAACTTGAGTGCAAGAGGGGAGAGTGGAATTCCATGTGTAGCGGTGAAATGCGTAGATATATGGAGGAACACCGGTGGCGAAAGCGGCTCTCTGGCTTGTAACTGACGCTGAGGCTCGAAAGCGTGGGGAGCAAACAGG     Streptococcaceae
#>                                                                                                                                                                                                                                                                          Genus
#> TACGGAGGGTGCAAGCGTTAATCGGAATTACTGGGCGTAAAGCGCACGCAGGCGGTCTGTCAAGTCGGATGTGAAATCCCCGGGCTCAACCTGGGAACTGCATTCGAAACTGGCAGGCTAGAGTCTTGTAGAGGGGGGTAGAATTCCAGGTGTAGCGGTGAAATGCGTAGAGATCTGGAGGAATACCGGTGGCGAAGGCGGCCCCCTGGACAAAGACTGACGCTCAGGTGCGAAAGCGTGGGGAGCAAACAGG       Klebsiella
#> TACGTAGGGTGCGAGCGTTGTCCGGAATTACTGGGCGTAAAGGGCTCGTAGGTGGTTTGTCGCGTCGTCTGTGAAATTCTGGGGCTTAACTCCGGGCGTGCAGGCGATACGGGCATAACTTGAGTGCTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCGAACAGG Corynebacterium
#> TACGTAGGGTGCAAGCGTTGTCCGGAATTACTGGGCGTAAAGAGCTCGTAGGTGGTTTGTCACGTCGTCTGTGAAATTCCACAGCTTAACTGTGGGCGTGCAGGCGATACGGGCTGACTTGAGTACTGTAGGGGTAACTGGAATTCCTGGTGTAGCGGTGAAATGCGCAGATATCAGGAGGAACACCGATGGCGAAGGCAGGTTACTGGGCAGTTACTGACGCTGAGGAGCGAAAGCATGGGTAGCAAACAGG  Corynebacterium
#> TACGTAGGTGACAAGCGTTGTCCGGATTTATTGGGCGTAAAGGGAGCGCAGGCGGTCTGTTTAGTCTAATGTGAAAGCCCACGGCTTAACCGTGGAACGGCATTGGAAACTGACAGACTTGAATGTAGAAGAGGAAAATGGAATTCCAAGTGTAGCGGTGGAATGCGTAGATATTTGGAGGAACACCAGTGGCGAAGGCGATTTTCTGGTCTAACATTGACGCTGAGGCTCGAAAGCGTGGGGAGCGAACAGG   Dolosigranulum
#> TACGTAGGTCCCGAGCGTTGTCCGGATTTATTGGGCGTAAAGCGAGCGCAGGCGGTTAGATAAGTCTGAAGTTAAAGGCTGTGGCTTAACCATAGTACGCTTTGGAAACTGTTTAACTTGAGTGCAAGAGGGGAGAGTGGAATTCCATGTGTAGCGGTGAAATGCGTAGATATATGGAGGAACACCGGTGGCGAAAGCGGCTCTCTGGCTTGTAACTGACGCTGAGGCTCGAAAGCGTGGGGAGCAAACAGG     Streptococcus
#> 
#> $metadata
#>                                                           LabID
#> SAMPLED_5080-MS-1_328-GATCTACG-TCGACGAG_S328_L001 participant01
#> SAMPLED_5080-MS-1_339-ACTCACTG-GATCGTGT_S339_L001 participant02
#> SAMPLED_5348-MS-1_162-ACGTGCGC-GGATATCT_S162_L001 participant03
#> SAMPLED_5348-MS-1_297-GTCTGCTA-ACGTCTCG_S297_L001 participant04
#> SAMPLED_5348-MS-1_381-TGCTCGTA-GTCAGATA_S381_L001  CTRL_pos_pcr
#> SAMPLED_5080-MS-1_307-ATAGTACC-ACGTCTCG_S307_L001  CTRL_neg_ext
#> SAMPLED_5080-MS-1_313-GACATAGT-TCGACGAG_S313_L001  CTRL_neg_pcr
#> FakeNeg1                                           CTRL_neg_ext
#> FakeNeg2                                           CTRL_neg_ext
#> FakeNeg3                                           CTRL_neg_ext
#> FakeNeg4                                           CTRL_neg_pcr
#> FakeNeg5                                           CTRL_neg_pcr
#> FakeNeg6                                           CTRL_neg_pcr
#> FakeNeg7                                           CTRL_neg_pcr
#> FakeNeg8                                           CTRL_neg_pcr
#>                                                         SampleType host_age
#> SAMPLED_5080-MS-1_328-GATCTACG-TCGACGAG_S328_L001     Homo sapiens       33
#> SAMPLED_5080-MS-1_339-ACTCACTG-GATCGTGT_S339_L001     Homo sapiens       25
#> SAMPLED_5348-MS-1_162-ACGTGCGC-GGATATCT_S162_L001     Homo sapiens       27
#> SAMPLED_5348-MS-1_297-GTCTGCTA-ACGTCTCG_S297_L001     Homo sapiens       26
#> SAMPLED_5348-MS-1_381-TGCTCGTA-GTCAGATA_S381_L001 positive control       NA
#> SAMPLED_5080-MS-1_307-ATAGTACC-ACGTCTCG_S307_L001 negative control       NA
#> SAMPLED_5080-MS-1_313-GACATAGT-TCGACGAG_S313_L001 negative control       NA
#> FakeNeg1                                          negative control       NA
#> FakeNeg2                                          negative control       NA
#> FakeNeg3                                          negative control       NA
#> FakeNeg4                                          negative control       NA
#> FakeNeg5                                          negative control       NA
#> FakeNeg6                                          negative control       NA
#> FakeNeg7                                          negative control       NA
#> FakeNeg8                                          negative control       NA
#>                                                   host_sex Host_disease   neg
#> SAMPLED_5080-MS-1_328-GATCTACG-TCGACGAG_S328_L001   female      healthy FALSE
#> SAMPLED_5080-MS-1_339-ACTCACTG-GATCGTGT_S339_L001     male      healthy FALSE
#> SAMPLED_5348-MS-1_162-ACGTGCGC-GGATATCT_S162_L001     male     COVID-19 FALSE
#> SAMPLED_5348-MS-1_297-GTCTGCTA-ACGTCTCG_S297_L001   female     COVID-19 FALSE
#> SAMPLED_5348-MS-1_381-TGCTCGTA-GTCAGATA_S381_L001     <NA>         <NA> FALSE
#> SAMPLED_5080-MS-1_307-ATAGTACC-ACGTCTCG_S307_L001     <NA>         <NA>  TRUE
#> SAMPLED_5080-MS-1_313-GACATAGT-TCGACGAG_S313_L001     <NA>         <NA>  TRUE
#> FakeNeg1                                              <NA>         <NA>  TRUE
#> FakeNeg2                                              <NA>         <NA>  TRUE
#> FakeNeg3                                              <NA>         <NA>  TRUE
#> FakeNeg4                                              <NA>         <NA>  TRUE
#> FakeNeg5                                              <NA>         <NA>  TRUE
#> FakeNeg6                                              <NA>         <NA>  TRUE
#> FakeNeg7                                              <NA>         <NA>  TRUE
#> FakeNeg8                                              <NA>         <NA>  TRUE
#>